+++ packages/core/src/api/meta/meta.test.ts
@@ -3,11 +3,11 @@ import { mockConfig } from '../../testUtils';
 
 describe('metaApi', () => {
   beforeAll(() => {
-    delete (global as any).FARO_BUILD_ID_FOO;
+    delete (global as any).FARO_BUNDLE_ID_FOO;
   });
 
   afterAll(() => {
-    delete (global as any).FARO_BUILD_ID_FOO;
+    delete (global as any).FARO_BUNDLE_ID_FOO;
   });
 
   it('can get the build ID from the global object', () => {
@@ -15,7 +15,7 @@ describe('metaApi', () => {
 
     expect(api.getBuildId('foo')).toBeUndefined();
 
-    (global as any).FARO_BUILD_ID_FOO = 'bar';
+    (global as any).FARO_BUNDLE_ID_FOO = 'bar';
     expect(api.getBuildId('foo')).toEqual('bar');
   });
 });