+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next
 
-- Enhancement (`@grafana/faro-core`): sourcemap uploads - add `buildId` to the `MetaApp` `Meta` object (#476).
+- Enhancement (`@grafana/faro-core`): sourcemap uploads - add `bundleId` to the `MetaApp` `Meta` object (#476).
 
 ## 1.3.7